+++ 3d_routines.h
@@ -13,9 +13,9 @@
 #define MAX_VERTICE_COUNT 1024
 
 #define PREPARE_3D_MESH(OBJECT_HANDLER, OBJECT_VERT_LIST, OBJECT_FACE_LIST, ZOOM_LEVEL, Z_DISTANCE, FLAG_CULL) \
-                  OBJECT_HANDLER.verts = (int const *)&OBJECT_VERT_LIST; \
+                  OBJECT_HANDLER.verts = (short const *)&OBJECT_VERT_LIST; \
                   OBJECT_HANDLER.nverts = VERT_COUNT(OBJECT_VERT_LIST); \
-                  OBJECT_HANDLER.faces = (int const *)&OBJECT_FACE_LIST; \
+                  OBJECT_HANDLER.faces = (short const *)&OBJECT_FACE_LIST; \
                   OBJECT_HANDLER.nfaces = FACE_COUNT(OBJECT_FACE_LIST); \
                   OBJECT_HANDLER.zoom = ZOOM_LEVEL; \
                   OBJECT_HANDLER.distance = Z_DISTANCE; \
@@ -35,19 +35,19 @@
 
 struct obj_3d
 {
-    int const* verts;
-    int nverts;
-    int const* faces;
-    int nfaces;
-    int zoom;
-    int distance;
-    int flag_cull_backfaces;
+    short const* verts;
+    short nverts;
+    short const* faces;
+    short nfaces;
+    short zoom;
+    short distance;
+    short flag_cull_backfaces;
 };
 
-void DrawAALine(int x1, int y1, int x2, int y2);
+void DrawAALine(short x1, short y1, short x2, short y2);
 
 void Prepare3DVertexList(void);
 
 void Delete3DVertexList(void);
 
-int Draw3DMesh(int rx, int ry, int y_offset, int m_scale_x);
+short Draw3DMesh(short rx, short ry, short y_offset, short m_scale_x);
