+++ Assets/fonts.h
@@ -1,4 +1,4 @@
 /* Font descriptor headers */
 
 extern const char future_font_glyph_array[88];
-extern const int future_font_x_pos_array[89];
+extern const short future_font_x_pos_array[89];
